test(e2e): extract startShare/stopShare utilities
