+++ apps/chat-bot/e2e/tests/character/share-with-learners-grace-window.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from '@playwright/test';
 import { AUTH_FILES } from '../../utils/const';
 import { configureCharacter, createCharacter } from '../../utils/character';
-import { waitForAutosave } from '../../utils/utils';
+import { startShare, stopShare, waitForAutosave } from '../../utils/utils';
 import { nanoid } from 'nanoid';
 
 test.describe('share character with grace window', () => {
@@ -22,18 +22,7 @@ test.describe('share character with grace window', () => {
 
   test('can extend a share session that has not yet expired', async ({ page }) => {
     await page.goto(`/characters/editor/${characterId}`);
-    await page.waitForURL(`**/characters/editor/${characterId}**`);
-
-    // Select usage time before starting share
-    await page.getByTestId('usage-time-select').click();
-    await page.getByTestId('usage-time-option-30').click();
-
-    await page.getByTestId('start-share-button').click();
-
-    await page.waitForURL('**/characters/editor/**/share');
-
-    // Navigate back to editor to verify share is active
-    await page.goto(`/characters/editor/${characterId}`);
+    await startShare(page, { id: characterId });
 
     await expect(page.getByTestId('add-additional-time-button')).toBeVisible();
     await expect(page.getByTestId('stop-share-button')).toBeVisible();
@@ -41,28 +30,8 @@ test.describe('share character with grace window', () => {
 
   test('can manually stop a share session', async ({ page }) => {
     await page.goto(`/characters/editor/${characterId}`);
-    await page.waitForURL(`**/characters/editor/${characterId}**`);
-
-    // Start sharing with 30 minutes
-    await page.getByTestId('usage-time-select').click();
-    await page.getByTestId('usage-time-option-30').click();
-    await page.getByTestId('start-share-button').click();
-
-    await page.waitForURL('**/characters/editor/**/share');
-
-    // Navigate back to editor
-    await page.goto(`/characters/editor/${characterId}`);
-    await page.waitForURL(`**/characters/editor/${characterId}**`);
-
-    const stopShareButton = page.getByTestId('stop-share-button');
-    await expect(stopShareButton).toBeVisible();
-    await stopShareButton.click();
-
-    const stopShareDialog = page.getByTestId('stop-share-dialog');
-    await expect(stopShareDialog).toBeVisible();
-
-    await page.getByTestId('stop-share-confirm-button').click();
-    await expect(stopShareDialog).not.toBeVisible();
+    await startShare(page, { id: characterId });
+    await stopShare(page);
 
     await expect(page.getByTestId('start-share-button')).toBeVisible();
   });
